requirements.txt
pillow
torchvision
torch
